requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.31.0
stringcase==1.2.0
python-dateutil==2.8.2
Flask-SQLAlchemy==2.5.1
